fix: next-auth types
