+++ admin-ui/middleware.ts
@@ -1,4 +1,3 @@
-import { ExtendedJWT } from "app/pages/api/auth/[...nextauth]";
 import { getToken } from "next-auth/jwt";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -11,10 +10,10 @@ export default async function middleware(req: NextRequest) {
   console.log("auth middleware");
   if (authEnabled) {
     console.log("auth enabled");
-    const token = (await getToken({
+    const token = await getToken({
       req,
       secret: nextAuthSecret,
-    })) as ExtendedJWT;
+    });
 
     if (!token?.user) {
       console.log("got user token");
