+++ admin-ui/src/modules/auth/axiosClient.ts
@@ -2,7 +2,6 @@ import axios from "axios";
 import applyCaseMiddleware from "axios-case-converter";
 import { getSession } from "next-auth/react";
 import { PROFILE_TOKEN_HEADER } from "app/common/const";
-import { ExtendedSession } from "../../pages/api/auth/[...nextauth]";
 
 const axiosOptions = {
   timeout: 20000,
@@ -18,7 +17,7 @@ const axiosClient = applyCaseMiddleware(axios.create(axiosOptions));
 
 if (typeof window !== "undefined" && authEnabled) {
   axiosClient.interceptors.request.use(async (req) => {
-    const session = (await getSession()) as ExtendedSession;
+    const session = await getSession();
 
     if (session?.apiTokens?.tilavaraus) {
       req.headers.Authorization = `Bearer ${session.apiTokens.tilavaraus}`;
