+++ admin-ui/types/next-auth.d.ts
@@ -1,0 +1,50 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import NextAuth from "next-auth";
+
+declare module "next-auth/jwt" {
+  interface JWT {
+    accessToken: string;
+    accessTokenExpires: number;
+    refreshToken: string;
+    user: User;
+    apiTokens: APITokens;
+    error?: string;
+  }
+  interface JwtParams {
+    token: JWT;
+    user: User;
+    account: TunnistamoAccount;
+  }
+}
+
+declare module "next-auth" {
+  type APITokens = {
+    tilavaraus: string;
+    profile: string;
+  };
+  type TunnistamoAccount = {
+    provider: string;
+    type: "oauth";
+    providerAccountId: string;
+    access_token: string;
+    refresh_token: string;
+    token_type: "bearer";
+    expires_at: number;
+    id_token: string;
+  };
+
+  interface User {
+    id: string;
+    name: string;
+    given_name: string;
+    family_name: string;
+    nickname: string;
+    email: string;
+    email_verified: boolean;
+  }
+  interface Session {
+    accessToken: string;
+    accessTokenExpires: number;
+    apiTokens: APITokens;
+  }
+}
